requirements.txt
serial
RPi.GPIO